requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
aenum=3.1.15=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=24.3.0=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py312h6a678d5_7
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
comm=0.2.2=pyhd8ed1ab_0
configargparse=1.7=pypi_0
contourpy=1.2.1=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py312h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
fonttools=4.51.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpcio=1.69.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.4=py312h06a4308_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.3=pyhd33586a_0
ipython=8.24.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=py312h06a4308_0
joblib=1.4.0=pypi_0
jpeg=9e=h5eee18b_1
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.5.0=py312h06a4308_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h77fa898_7
libgomp=13.2.0=h77fa898_7
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lightly=1.5.17=pypi_0
lightly-utils=0.0.2=pypi_0
lightning-utilities=0.11.9=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.7=pypi_0
markupsafe=2.1.3=py312h5eee18b_0
matplotlib=3.8.4=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
mpmath=1.3.0=py312h06a4308_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py312h06a4308_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
omegaconf=2.3.0=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.3.0=h4ab18f5_3
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=py312h5eee18b_0
pip=23.3.1=py312h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
propcache=0.2.1=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pydantic=2.10.5=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.2=h996f2a0_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.2.2=py3.12_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-lightning=2.5.0.post0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py312h5eee18b_0
pyzmq=25.1.2=py312h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.31.0=py312h06a4308_1
scikit-learn=1.4.2=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
sentry-sdk=2.18.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py312h06a4308_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.4.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.2.2=py312_cu118
torchmetrics=1.6.1=pypi_0
torchvision=0.17.2=py312_cu118
tornado=6.3.3=py312h5eee18b_0
tqdm=4.66.2=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.1.0=py312h06a4308_1
wandb=0.18.5=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.1.3=pypi_0
wheel=0.41.2=py312h06a4308_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.18.3=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
